requirement.txt
streamlit
pandas
python-dotenv
langchain
langchain-community
langchain-groq
sqlalchemy
